+++ eliza/apps/app-steward/src/index.ts
@@ -3,3 +3,4 @@ export * from "./ApprovalQueue";
 export * from "./TransactionHistory";
 export * from "./StewardLogo";
 export * from "./chain-utils";
+export * from "./services/steward-sidecar";