+++ eliza/apps/app-vincent/src/index.ts
@@ -2,6 +2,7 @@ export * from "./VincentAppView";
 export * from "./VincentConnectionCard";
 export * from "./TradingProfileCard";
 export * from "./TradingStrategyPanel";
+export * from "./routes";
 export * from "./useVincentState";
 export * from "./VaultStatusCard";
 export * from "./useVincentDashboard";